+++ EnforcerMod_VS/Modules/Buffs.cs
@@ -14,6 +14,7 @@ namespace Modules {
         internal static BuffDef energyShieldBuff;
         internal static BuffDef skateboardBuff;
         internal static BuffDef minigunBuff;
+        internal static BuffDef nemforcerRegenBuff;
 
         internal static BuffDef smallSlowBuff;
         internal static BuffDef bigSlowBuff;
@@ -26,7 +27,8 @@ namespace Modules {
             energyShieldBuff = AddNewBuff("EnergyShield", Assets.MainAssetBundle.LoadAsset<Sprite>("texBuffProtectAndServe"), EnforcerModPlugin.characterColor, false, false);
             skateboardBuff = AddNewBuff("Swag", Assets.LoadBuffSprite("RoR2/Base/SprintOutOfCombat/bdWhipBoost.asset"), EnforcerModPlugin.characterColor, false, false);
             minigunBuff = AddNewBuff("MinigunStance", Assets.MainAssetBundle.LoadAsset<Sprite>("texBuffMinigun"), new Color(1, 0.7176f, 0.1725f), false, false);
-                                                      
+            nemforcerRegenBuff = AddNewBuff("NemforcerRegen", Addressables.LoadAssetAsync<BuffDef>("RoR2/Base/Croco/bdCrocoRegen.asset").WaitForCompletion().iconSprite, new Color32(210, 50, 22, 255), true, false);
+
             Sprite slowIcon =  Assets.LoadBuffSprite("RoR2/Base/Common/bdSlow50.asset");
             Sprite cloakIcon = Assets.LoadBuffSprite("RoR2/Base/Common/bdCloak.asset");