+++ EnforcerMod_VS/Modules/Tokens.cs
@@ -2,6 +2,7 @@
 using EntityStates.Enforcer;
 using EntityStates.Enforcer.NeutralSpecial;
 using EntityStates.Nemforcer;
+using R2API;
 using System;
 
 namespace Modules {
@@ -283,6 +284,7 @@ namespace Modules {
             //passive
             LanguageAPI.Add("NEMFORCER_PASSIVE_NAME", "Colossus");
             LanguageAPI.Add("NEMFORCER_PASSIVE_DESCRIPTION", $"Nemesis Enforcer gains <style=cIsHealing>bonus health regen</style>, based on his current <style=cIsHealth>missing health</style>, up to <style=cIsHealth>{100 * NemforcerPlugin.passiveRegenBonus}% max health</style>.");
+            LanguageAPI.Add("NEMFORCER_PASSIVE_REWORK_DESCRIPTION", $"Nemesis Enforcer regenerates <style=cIsHealing>{NemforcerPlugin.nemforcerRegenBuffPercent}%</style> of his <style=cIsHealing>maximum health</style> over {NemforcerPlugin.nemforcerRegenBuffDuration}s when striking enemies with <style=cIsDamage>melee attacks</style>.");
 
             #region Primary
             LanguageAPI.Add("NEMFORCER_PRIMARY_HAMMER_NAME", "Golden Hammer");